+++ objectified-ui/src/app/layout.tsx
@@ -3,6 +3,7 @@ import "./globals.css";
 import "@radix-ui/themes/styles.css";
 import SessionWrapper from "@/app/components/auth/SessionWrapper";
 import ThemeRegistry from "@/app/components/theme/ThemeRegistry";
+import { ThemeProvider } from "@/app/providers/ThemeProvider";
 import { DialogProvider } from "@/app/components/providers/DialogProvider";
 import { Toaster } from "@/app/components/ui/Toaster";
 import { ThemeProvider as NextThemesProvider } from "next-themes";
@@ -36,10 +37,12 @@ export default function RootLayout({
           >
             <ThemeRegistry>
               <SessionWrapper>
-                <DialogProvider>
-                  {children}
-                  <Toaster />
-                </DialogProvider>
+                <ThemeProvider>
+                  <DialogProvider>
+                    {children}
+                    <Toaster />
+                  </DialogProvider>
+                </ThemeProvider>
               </SessionWrapper>
             </ThemeRegistry>
           </RadixTheme>